requirements.txt
bottle
GitPython